Fix passwords not working
